Commit the SAP HANA Cloud Samples - Welcome application
This is the initial version of the "Welcome" application which is
pre-deployed in every new trial account and allows you to get hands-on
experience with the SAP HANA Cloud capabilities.
It comprises three demo applications - HANA demo, Paul the Octopus and
SAP Library.
By playing around with the Welcome application, you will learn how to
do the following:

** Manage your applications from the cockpit, that is starting and
stopping an application
** Check logs
** View application processes and their metrics
** Check performance

Change-Id: I3f39ab4e862a172898f6920df028aa8e87319d7b